updated rating method

+++ server.py
@@ -219,7 +219,7 @@ def update_location():
 
             location_rating = get_country_rating(latitude, longitude, world)
 
-            overall_rating = location_rating*((1+time_difference_seconds)**(100-location_rating))
+            overall_rating = int(location_rating*((1+time_difference_seconds)**(100-location_rating)))
 
             # Notify clients to refresh the map
             socketio.emit('refresh_map', {'message': 'New location received'})
